Move the UUIDModelMixin in core.mixins
